requirements.txt
Django==2.2.10
django-extensions==2.2.9
django-filter==2.2.0
django-shell-plus==1.1.7
djangorestframework==3.11.0
Markdown==3.1.1
mccabe==0.6.1
Pillow==7.1.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
pylama==7.7.1
pytz==2019.3
six==1.14.0
snowballstemmer==2.0.0
sqlparse==0.3.1